python/requirements.txt
python-dotenv
langchain-openai
requests
ollama
jinja2
fastapi
openai==1.35.15
uvicorn[standard]